+++ iris-mock/src/main/kotlin/dev/arildo/irismock/IrisMockWrapper.kt
@@ -2,13 +2,12 @@
 
 package dev.arildo.irismock
 
-import dev.arildo.irismock.util.log
 import okhttp3.Interceptor
 import okhttp3.Response
 
 /*
- * When using the IrisMock plugin to inject the interceptors,
- * this class is called to iterate over them, granting all logic is processed.
+ * When using IrisMock plugin to inject interceptors, this class is called
+ * to iterate over them, granting all modifiers are processed.
  *
  * When using only the IrisMock runtime dependency, this class is not called at all,
  * since the interceptor is being added manually to the OkHttp builder.
@@ -17,10 +16,6 @@ internal class IrisMockWrapper : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
         IrisMock.interceptors.forEach { it().intercept(chain) }
 
-        val response = ModifierProcessor.process(chain)
-
-        if (IrisMock.enableLogs) response.log()
-
-        return response
+        return ModifierProcessor.process(chain)
     }
 }
