--- iris-mock/src/main/kotlin/dev/arildo/irismock/util/Constants.kt
@@ -1,6 +1,0 @@
-package dev.arildo.irismock.util
-
-internal const val IRIS_MOCK_TAG = "IrisMock"
-internal const val APPLICATION_JSON = "application/json"
-internal const val IRIS_MOCK_INTERCEPTORS = "dev.arildo.irismock.IrisMockContainer"
-internal const val INTERCEPTORS_FIELD = "interceptors"
