+++ iris-mock/src/main/kotlin/dev/arildo/irismock/util/Utils.kt
@@ -1,15 +1,14 @@
 package dev.arildo.irismock.util
 
 import dev.arildo.irismock.IrisMock
-import okhttp3.Interceptor.Chain
-import okhttp3.MediaType
 import okhttp3.Protocol
 import okhttp3.Request
 import okhttp3.Response
-import okhttp3.ResponseBody
 import okio.Buffer
 import okio.GzipSource
 import okio.buffer
+
+const val APPLICATION_JSON = "application/json"
 
 internal fun Response.log() {
     val body = getResponseBody(this)
@@ -53,18 +52,3 @@ private fun getResponseBody(response: Response): String {
 
     return body
 }
-
-// since [IRIS_MOCK_INTERCEPTORS] is generated by the IrisMock plugin,
-// we can infer the response from the existence of the generated class itself
-fun isUsingIrisMockPlugin(): Boolean = runCatching {
-    Class.forName(IRIS_MOCK_INTERCEPTORS) ?: throw ClassNotFoundException()
-    true
-}.getOrDefault(false)
-
-fun createBlankResponse(chain: Chain): Response = Response.Builder()
-    .request(chain.request())
-    .protocol(Protocol.HTTP_1_1)
-    .code(HttpCode.OK.code)
-    .body(ResponseBody.create(MediaType.get(APPLICATION_JSON), ""))
-    .message("")
-    .build()
