+++ iris-mock/src/test/kotlin/dev/arildo/irismock/dsl/RequestMethodDslTest.kt
@@ -1,10 +1,10 @@
 package dev.arildo.irismock.dsl
 
 import dev.arildo.irismock.ModifierProcessor
+import dev.arildo.irismock.Utils.createBlankResponse
 import dev.arildo.irismock.callmodifier.CustomResponseBodyModifier
 import dev.arildo.irismock.util.APPLICATION_JSON
 import dev.arildo.irismock.util.Method
-import dev.arildo.irismock.util.createBlankResponse
 import io.mockk.every
 import io.mockk.mockk
 import okhttp3.Interceptor
